Moved all well index calculation functions into Reservoir lib
